+++ apps/backend/src/libs/constants/constants.ts
@@ -9,3 +9,4 @@ export {
 	NO_ROWS_AFFECTED,
 } from "./repository.constant.js";
 export { WHITE_ROUTES } from "./white-routes.constant.js";
+export { EMPTY_ARRAY_LENGTH } from "@meetlytic/shared";