+++ apps/backend/src/libs/modules/server-application/base-server-application.ts
@@ -24,7 +24,7 @@ import { HTTPCode, HTTPError } from "~/libs/modules/http/http.js";
 import { type Logger } from "~/libs/modules/logger/logger.js";
 import { type BaseSocketService } from "~/libs/modules/socket/socket.js";
 import { jwt } from "~/libs/modules/token/token.js";
-import { authorizationPlugin } from "~/libs/plugins/authorization/authorization.plugin.js";
+import { authorization } from "~/libs/plugins/authorization/authorization.plugin.js";
 import { uploadPlugin } from "~/libs/plugins/uploads/upload.plugin.js";
 import {
 	type ServerCommonErrorResponse,
@@ -155,7 +155,7 @@ class BaseServerApplication implements ServerApplication {
 			`multipart parser: ${String(this.app.hasContentTypeParser("multipart"))}`,
 		);
 
-		await this.app.register(authorizationPlugin, {
+		await this.app.register(authorization, {
 			routesWhiteList: WHITE_ROUTES,
 			services: {
 				config: this.config,
@@ -210,7 +210,7 @@ class BaseServerApplication implements ServerApplication {
 		this.app.route({
 			handler,
 			method,
-			...(preHandlers?.length ? { preHandler: preHandlers } : {}),
+			preHandler: preHandlers ?? [],
 			schema,
 			url: path,
 		});