+++ apps/backend/src/modules/meetings/meetings.service.ts
@@ -1,3 +1,4 @@
+import { EMPTY_ARRAY_LENGTH } from "~/libs/constants/constants.js";
 import { APIPath } from "~/libs/enums/enums.js";
 import { AuthError } from "~/libs/exceptions/exceptions.js";
 import {
@@ -6,6 +7,11 @@ import {
 } from "~/libs/modules/cloud-formation/cloud-formation.js";
 import template from "~/libs/modules/cloud-formation/libs/templates/ec2-instance-template.json" with { type: "json" };
 import { HTTPCode } from "~/libs/modules/http/http.js";
+import {
+	SocketEvent,
+	SocketNamespace,
+} from "~/libs/modules/socket/libs/enums/enums.js";
+import { socketService } from "~/libs/modules/socket/socket.js";
 import {
 	type BaseToken,
 	type SharedJwtPayload,
@@ -21,6 +27,7 @@ import {
 	type MeetingGetAllResponseDto,
 	type MeetingGetPublicUrlResponseDto,
 	type MeetingResponseDto,
+	type MeetingTranscriptionGetAllResponseDto,
 	type MeetingTranscriptionRequestDto,
 	type MeetingTranscriptionResponseDto,
 	type MeetingUpdateRequestDto,
@@ -85,6 +92,18 @@ class MeetingService implements Service<MeetingResponseDto> {
 			throw new MeetingError({
 				message: MeetingErrorMessage.INVALID_MEETING_LINK,
 				status: HTTPCode.BAD_REQUEST,
+			});
+		}
+
+		const existing = await this.meetingRepository.findAll({
+			meetingId: meetingId,
+			status: MeetingStatus.STARTED,
+		});
+
+		if (existing.length > EMPTY_ARRAY_LENGTH) {
+			throw new MeetingError({
+				message: MeetingErrorMessage.DUPLICATED_MEETING,
+				status: HTTPCode.CONFLICT,
 			});
 		}
 
@@ -157,7 +176,6 @@ class MeetingService implements Service<MeetingResponseDto> {
 	}
 
 	public async endMeeting(id: number): Promise<MeetingDetailedResponseDto> {
-		await this.cloudFormation.delete(id);
 		const meeting = await this.meetingRepository.update(id, {
 			instanceId: null,
 			status: MeetingStatus.ENDED,
@@ -169,6 +187,8 @@ class MeetingService implements Service<MeetingResponseDto> {
 				status: HTTPCode.NOT_FOUND,
 			});
 		}
+
+		void this.cloudFormation.delete(id);
 
 		return meeting.toDetailedObject();
 	}
@@ -224,6 +244,12 @@ class MeetingService implements Service<MeetingResponseDto> {
 		};
 	}
 
+	public async getTranscriptionsByMeetingId(
+		id: number,
+	): Promise<MeetingTranscriptionGetAllResponseDto> {
+		return await this.meetingTranscriptionService.getByMeetingId(id);
+	}
+
 	public async saveChunk({
 		chunkText,
 		meetingId,
@@ -237,10 +263,12 @@ class MeetingService implements Service<MeetingResponseDto> {
 	}
 
 	public async stopRecording(id: number): Promise<void> {
-		// TODO:
-		// 1. emit a message for the bot (bot stops audio recording, transcribes full audio, gets summary and action points)
-		// 2. move endMeeting(id) call to the websocket event handler
-		await this.endMeeting(id);
+		const meeting = await this.find(id);
+		socketService.emitTo({
+			event: SocketEvent.STOP_RECORDING,
+			namespace: SocketNamespace.BOTS,
+			room: String(meeting.id),
+		});
 	}
 
 	public async update(
@@ -257,21 +285,22 @@ class MeetingService implements Service<MeetingResponseDto> {
 		}
 
 		const meeting = MeetingEntity.initialize({
-			actionItems: meetingEntity.toDetailedObject().actionItems,
+			actionItems:
+				payload.actionItems ?? meetingEntity.toDetailedObject().actionItems,
 			createdAt: meetingEntity.toObject().createdAt,
-			host: payload.host,
+			host: payload.host ?? meetingEntity.toObject().host,
 			id,
 			instanceId: meetingEntity.toObject().instanceId,
 			meetingId: meetingEntity.toObject().meetingId,
 			meetingPassword: meetingEntity.toObject().meetingPassword,
 			ownerId: meetingEntity.toObject().ownerId,
-			status: payload.status,
-			summary: meetingEntity.toDetailedObject().summary,
+			status: payload.status ?? meetingEntity.toObject().status,
+			summary: payload.summary ?? meetingEntity.toDetailedObject().summary,
 		});
 
 		const updatedMeeting = await this.meetingRepository.update(
 			id,
-			meeting.toNewObject(),
+			meeting.toDetailedObject(),
 		);
 
 		if (!updatedMeeting) {