+++ apps/bot/src/libs/modules/socket-client/base-socket.module.ts
@@ -1,5 +1,7 @@
 import { io, type Socket } from "socket.io-client";
 
+import { logger } from "../logger/logger.js";
+import { SocketNamespace, Timeout } from "./enums/enums.js";
 import {
 	type ClientToServerEvents,
 	type ServerToClientEvents,
@@ -21,14 +23,21 @@ class BaseSocketClient {
 		this.url = url;
 	}
 
+	private getOrigin = (url: string, namespace: string): string =>
+		`${url}${namespace}`;
+
 	public connect(): void {
 		if (!this.socket) {
-			this.socket = io(this.url, {
+			this.socket = io(this.getOrigin(this.url, SocketNamespace.BOTS), {
 				reconnection: true,
 				reconnectionAttempts: Infinity,
-				reconnectionDelay: 1000,
+				reconnectionDelay: Timeout.ONE_SECOND,
 				transports: ["websocket"],
 			});
+
+			logger.info(
+				`Bot is connecting to  ${this.getOrigin(this.url, SocketNamespace.BOTS)}`,
+			);
 		} else if (!this.socket.connected) {
 			this.socket.connect();
 		}
@@ -43,6 +52,13 @@ class BaseSocketClient {
 		...parameters: Parameters<ClientToServerEvents[K]>
 	): void {
 		this.socket?.emit(event, ...parameters);
+	}
+
+	public off<
+		K extends keyof ServerToClientEvents,
+		L extends keyof SocketReservedEvents,
+	>(event: K | L, listener: Listener<K | L>): void {
+		this.socket?.off(event, listener as Listener<K>);
 	}
 
 	public on<