+++ apps/bot/src/libs/modules/socket-client/libs/types/client-to-server-events.type.ts
@@ -1,6 +1,12 @@
-import { type MeetingTranscriptionRequestDto } from "~/libs/types/types.js";
+import {
+	type MeetingSummaryActionItemsResponseDto,
+	type MeetingTranscriptionRequestDto,
+} from "~/libs/types/types.js";
 
 type ClientToServerEvents = {
+	joinRoom: (meetingId: string) => void;
+	recordingStopped: (meetingId: string) => void;
+	saveSummaryActionItems: (data: MeetingSummaryActionItemsResponseDto) => void;
 	transcribe: (data: MeetingTranscriptionRequestDto) => void;
 };