+++ apps/frontend/src/modules/users/libs/default-values/sign-in.default-values.ts
@@ -1,8 +1,0 @@
-import { type UserSignInRequestDto } from "../../users.js";
-
-const DEFAULT_SIGN_IN_VALUES: UserSignInRequestDto = {
-	email: "",
-	password: "",
-};
-
-export { DEFAULT_SIGN_IN_VALUES };