+++ apps/frontend/src/modules/users/slices/users.slice.ts
@@ -2,6 +2,7 @@ import { createSlice, type PayloadAction } from "@reduxjs/toolkit";
 
 import { DataStatus } from "~/libs/enums/enums.js";
 import { type UserWithDetailsDto, type ValueOf } from "~/libs/types/types.js";
+import { logout } from "~/modules/auth/slices/actions.js";
 
 import { getProfile, updateProfile } from "./actions.js";
 import {
@@ -56,6 +57,10 @@ const { actions, reducer } = createSlice({
 		});
 		builder.addCase(updateProfile.rejected, (state) => {
 			state.dataStatus = DataStatus.REJECTED;
+		});
+		builder.addCase(logout.fulfilled, (state) => {
+			state.user = null;
+			state.dataStatus = DataStatus.IDLE;
 		});
 
 		builder.addCase(uploadAvatar.pending, (state) => {