+++ apps/frontend/src/pages/profile/profile.tsx
@@ -10,6 +10,7 @@ import {
 	useCallback,
 	useEffect,
 } from "~/libs/hooks/hooks.js";
+import { StorageKey } from "~/libs/modules/storage/storage.js";
 import {
 	actions as userActions,
 	type UserUpdateResponseDto,
@@ -19,16 +20,16 @@ import { ProfileForm } from "./components/components.js";
 import styles from "./styles.module.css";
 
 const Profile: React.FC = () => {
+	const token = localStorage.getItem(StorageKey.TOKEN);
 	const dispatch = useAppDispatch();
 	const user = useAppSelector((state) => state.users.user);
-
 	const status = useAppSelector((state) => state.users.dataStatus);
 
 	useEffect(() => {
-		if (status === DataStatus.IDLE) {
+		if (status === DataStatus.IDLE && token) {
 			void dispatch(userActions.getProfile());
 		}
-	}, [status, dispatch]);
+	}, [token, status, dispatch]);
 
 	const handleSubmit = useCallback(
 		(payload: UserUpdateResponseDto): void => {