+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## [1.9.1](https://github.com/BinaryStudioAcademy/bsa-2025-meetlytic/compare/backend-v1.9.0...backend-v1.9.1) (2025-08-20)
+
+
+### Bug Fixes
+
+* change error code for delete and get meeting ml-222 ([#234](https://github.com/BinaryStudioAcademy/bsa-2025-meetlytic/issues/234)) ([23e57ad](https://github.com/BinaryStudioAcademy/bsa-2025-meetlytic/commit/23e57adde62728ad789213b13297c1acd54e82ba))
+
+## [1.9.0](https://github.com/BinaryStudioAcademy/bsa-2025-meetlytic/compare/backend-v1.8.0...backend-v1.9.0) (2025-08-19)
+
+
+### Features
+
+* add a possibility to specify instance type from env file ml-236 ([#239](https://github.com/BinaryStudioAcademy/bsa-2025-meetlytic/issues/239)) ([4f8c690](https://github.com/BinaryStudioAcademy/bsa-2025-meetlytic/commit/4f8c690ee04eb5c419258fda7fc9f3b88ba90197))
+
+## [1.8.0](https://github.com/BinaryStudioAcademy/bsa-2025-meetlytic/compare/backend-v1.7.0...backend-v1.8.0) (2025-08-19)
+
+
+### Features
+
+* add meeting details screen ml-131 ([#191](https://github.com/BinaryStudioAcademy/bsa-2025-meetlytic/issues/191)) ([1dccbb7](https://github.com/BinaryStudioAcademy/bsa-2025-meetlytic/commit/1dccbb7f3b4ca85da0533931f54d51d83d5ac604))
+* send transcription to frontend real time ml-180 ([#214](https://github.com/BinaryStudioAcademy/bsa-2025-meetlytic/issues/214)) ([5fc5ee5](https://github.com/BinaryStudioAcademy/bsa-2025-meetlytic/commit/5fc5ee53cd0d840b3376a5686c8919881de3ffd1))
+
+
+### Bug Fixes
+
+* allow reconnection zoom meeting ml-230 ([#233](https://github.com/BinaryStudioAcademy/bsa-2025-meetlytic/issues/233)) ([b724890](https://github.com/BinaryStudioAcademy/bsa-2025-meetlytic/commit/b724890abda6e9703860a3345e7eb9052220edcd))
+* creating duplicate meeting after cloudFormation error ml-175 ([#224](https://github.com/BinaryStudioAcademy/bsa-2025-meetlytic/issues/224)) ([ed351ab](https://github.com/BinaryStudioAcademy/bsa-2025-meetlytic/commit/ed351abfa93a24651c7eec5982e9831cfa7da4c7))
+
 ## [1.7.0](https://github.com/BinaryStudioAcademy/bsa-2025-meetlytic/compare/backend-v1.6.0...backend-v1.7.0) (2025-08-15)